Reinforce user register
