--- src/main/java/com/bupt/eights/Controller/RoomController.java
@@ -1,5 +1,0 @@
-package com.bupt.eights.Controller;
-
-public class RoomController {
-
-}
