+++ src/main/java/com/bupt/eights/controller/AdminController.java
@@ -1,4 +1,4 @@
-package com.bupt.eights.Controller;
+package com.bupt.eights.controller;
 
 import lombok.extern.slf4j.Slf4j;
