+++ src/main/java/com/bupt/eights/controller/RoomController.java
@@ -1,0 +1,5 @@
+package com.bupt.eights.controller;
+
+public class RoomController {
+
+}
