+++ src/main/java/com/bupt/eights/utils/URLConstant.java
@@ -1,6 +1,6 @@
 package com.bupt.eights.utils;
 
-public class Constant {
+public class URLConstant {
     
     public static final String LOGIN_URL = "/api/v1/users";
