+++ src/main/resources/com/bupt/eights/mappers/AppointmentMapper.xml
@@ -1,14 +1,16 @@
-<!-- resources/mappers/AppointmentMapper.xml -->
-
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper
+    PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+    "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="com.bupt.eights.dao.AppointmentMapper">
     <select id="findOne" resultType="Appointment">
         SELECT a.*
         FROM Appointment a
-        WHERE id = #{appointmentId}
+        WHERE appointmentId = #{appointmentId}
     </select>
     
     <insert id="save" parameterType="Appointment">
-        INSERT INTO Appointment(`DATE`, `TIME`, NOTES, STATUS, STUDENT_NAME, ROOM_NAME)
-        VALUES(#{date}, #{time}, #{notes}, #{appointmentStatus}, #{student.name}, #{room.name});
+        INSERT INTO Appointment(date, time, notes, status, studentName, roomName)
+        VALUES(#{date}, #{time}, #{notes}, #{appointmentStatus, typeHandler=org.apache.ibatis.type.EnumTypeHandler}, #{student.name}, #{room.name});
     </insert>
 </mapper>
