+++ src/main/resources/com/bupt/eights/mappers/StudentMapper.xml
@@ -1,5 +1,7 @@
-<!-- resources/mappers/StudentMapper.xml -->
-
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper
+    PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+    "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="com.bupt.eights.dao.StudentMapper">
     
     <select id="findStudentByPhoneNumber" resultType="Student">
